simplification.py: Add average speed and total travel time variables
When simplification.py is ran, all attributes, including speed and travel time are currently copied as either a single value or list of unique values. While this is useful for most parameters, for speed and travel time this is not ideal, since the total travel time and average allowed speed can't be calculated anymore.

This commit adds calculation for the total travel time and average allowed speed to resolve this issue.